# internal/service/api/swagger/docs.yaml
basePath: /api/v1
definitions:
...
  routes.ErrorResponse:
    description: Standardized error response
    properties:
      error:
        description: Error message.
        example: Bad Request
        type: string
      stackTrace:
        description: Stack trace, only populated in debug mode.
        type: string
    type: object
...
  routes.MarkNotificationsViewedRequestJson:
    properties:
      ids:
        example:
        - ntf_test550e8400abcde
        items:
          type: string
        type: array
    type: object
...
  routes.OpenAPIListNotificationsResponseJson:
    description: Paginated list of actor-visible notifications
    properties:
      cursor:
        type: string
      items:
        items: {}
        type: array
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: AuthProxy is an open-source, embeddable integration platform-as-a-service
    (iPaaS). It manages the connection lifecycle to 3rd party systems, allowing applications
    to call those systems through an authenticating proxy.
  title: AuthProxy API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: List active actor-visible notifications
      parameters:
      - description: Maximum number of notifications to return
        in: query
        name: limit
        type: integer
      - description: Include notifications the actor has already viewed
        in: query
        name: includeViewed
        type: boolean
      - description: Notification state; defaults to active
        in: query
        name: state
        type: string
      - description: Filter by namespace
        in: query
        name: namespace
        type: string
      - description: Filter by denormalized resource label selector
        in: query
        name: labelSelector
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.OpenAPIListNotificationsResponseJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/_viewed:
    post:
      consumes:
      - application/json
      description: Mark multiple notifications viewed for the authenticated actor
      parameters:
      - description: Notification IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.MarkNotificationsViewedRequestJson'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notifications viewed
      tags:
      - notifications
  /notifications/{id}/_viewed:
    post:
      description: Mark a notification viewed for the authenticated actor
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification viewed
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: Bearer <token>)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"